pyproject.toml
[project]
name = "qlib-quant"
version = "0.1.0"
description = "A-share multi-factor quantitative investment system"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "sxt", email = "sxt@example.com"}
]
dependencies = [
    "pyqlib>=0.9",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "black>=23.0",
    "mypy>=1.0",
]
backtest = [
    "pybroker>=0.22",
]
full = [
    "pybroker>=0.22",
    "akshare>=1.12",
    "tushare>=1.3",
    "quantstats>=0.0.32",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "modules*", "utils*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
